help

+++ libft/ft_uitoa.c
@@ -6,13 +6,13 @@
 /*   By: dcastro- <dcastro-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/07/27 14:30:29 by dcastro-          #+#    #+#             */
-/*   Updated: 2017/08/16 17:15:34 by dcastro-         ###   ########.fr       */
+/*   Updated: 2017/08/16 19:33:33 by dcastro-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/libft.h"
 
-static unsigned int	num_len(unsigned int nbr)
+static unsigned int	num_len(uintmax_t nbr)
 {
 	unsigned int	i;
